+++ src/test/_setup.ts
@@ -1,6 +1,11 @@
 /*---------------------------------------------------------
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
+
+// Uncomment for verbose GRPC traces:
+// process.env.GRPC_VERBOSITY = 'DEBUG';
+// process.env.GRPC_TRACE = 'all';
+
 import * as chai from 'chai';
 import { ConnectionPool } from '../connection-pool';